[Ash] Feat: 이미지 별도 저장 순서 변경
